+++ mysql-client/src/test/scala/com/twitter/finatra/mysql/client/tests/EmbeddedMysqlServerIntegrationTest.scala
@@ -9,7 +9,7 @@ import java.nio.file.Paths
 class EmbeddedMysqlServerIntegrationTest extends Test with IgnoreMysqlHarnessIfUnavailable {
 
   test("EmbeddedMysqlServer#can start", MysqlHarnessTag) {
-    val server = EmbeddedMysqlServer()
+    val server = EmbeddedMysqlServer.newBuilder().withLazyStart.newServer()
     try {
       server.start()
       server.assertStarted()
@@ -19,8 +19,18 @@ class EmbeddedMysqlServerIntegrationTest extends Test with IgnoreMysqlHarnessIfU
     }
   }
 
+  test("EmbeddedMysqlServer#can start eagerly", MysqlHarnessTag) {
+    val server = EmbeddedMysqlServer.newBuilder().newServer()
+    try {
+      server.assertStarted()
+      server.close()
+    } finally {
+      server.close()
+    }
+  }
+
   test("EmbeddedMysqlServer#can retrieve a client", MysqlHarnessTag) {
-    val server = EmbeddedMysqlServer()
+    val server = EmbeddedMysqlServer.newBuilder().withLazyStart.newServer()
     try {
       val clnt = server.mysqlClient
       server.assertStarted()
@@ -32,7 +42,7 @@ class EmbeddedMysqlServerIntegrationTest extends Test with IgnoreMysqlHarnessIfU
   }
 
   test("EmbeddedMysqlServer#retrieving a client always creates a new instance", MysqlHarnessTag) {
-    val server = EmbeddedMysqlServer()
+    val server = EmbeddedMysqlServer.newBuilder().withLazyStart.newServer()
     try {
       val clnt = server.mysqlClient
       server.assertStarted()
@@ -57,6 +67,8 @@ class EmbeddedMysqlServerIntegrationTest extends Test with IgnoreMysqlHarnessIfU
     val user3 = User("user3", Some("pass3"), User.Permission.Select)
     val server = EmbeddedMysqlServer.newBuilder
       .withUsers(Seq(user1, user2, user3))
+      .withDatabaseName("dbWithMoreUsers")
+      .withLazyStart
       .newServer()
 
     try {
@@ -75,7 +87,10 @@ class EmbeddedMysqlServerIntegrationTest extends Test with IgnoreMysqlHarnessIfU
 
   test("EmbeddedMysqlServer#fails to start with bad config") {
     val server =
-      EmbeddedMysqlServer.newBuilder.withPath(Paths.get("/tmp/mysql/not_extracted")).newServer()
+      EmbeddedMysqlServer.newBuilder
+        .withPath(Paths.get("/tmp/mysql/not_extracted"))
+        .withLazyStart
+        .newServer()
     try {
       intercept[Exception] {
         server.start()
@@ -87,7 +102,7 @@ class EmbeddedMysqlServerIntegrationTest extends Test with IgnoreMysqlHarnessIfU
   }
 
   test("EmbeddedMysqlServer#cannot start a closed server", MysqlHarnessTag) {
-    val server = EmbeddedMysqlServer()
+    val server = EmbeddedMysqlServer.newBuilder().withLazyStart.newServer()
     try {
       server.start()
       server.assertStarted()