+++ mysql-client/src/test/scala/com/twitter/finatra/mysql/client/tests/MysqlClientModuleTraitTest.scala
@@ -99,7 +99,7 @@ class MysqlClientModuleTraitTest extends Test with Mockito with IgnoreMysqlHarne
   }
 
   test("MysqlClientModuleTrait#can bind over MysqlClient", MysqlHarnessTag) {
-    val embeddedMysql = EmbeddedMysqlServer()
+    val embeddedMysql = EmbeddedMysqlServer.newBuilder().withLazyStart.newServer()
     val mysqlClient = embeddedMysql.mysqlClient
     val server = new EmbeddedTwitterServer(
       new TwitterServer {